(refactor) updated pre-check-in tests
