+++ src/applications/check-in/pre-check-in/tests/e2e/errors/get-pre-check-in/error.in.body.cypress.spec.js
@@ -1,4 +1,4 @@
-import '../../support/commands';
+import '../../../../../tests/e2e/commands';
 
 import ApiInitializer from '../../../../../api/local-mock-api/e2e/ApiInitializer';
 import ValidateVeteran from '../../../../../tests/e2e/pages/ValidateVeteran';
