+++ src/applications/check-in/pre-check-in/tests/e2e/loading-session/returning.session.cypress.spec.js
@@ -1,4 +1,4 @@
-import '../support/commands';
+import '../../../../tests/e2e/commands';
 
 import ApiInitializer from '../../../../api/local-mock-api/e2e/ApiInitializer';
 import Introduction from '../../../tests/e2e/pages/Introduction';
